13-я проектная работа, 1-я итерация, 3-е исправление

+++ controllers/cards.js
@@ -1,3 +1,4 @@
+const mongoose = require('mongoose');
 const Card = require('../models/card');
 const { ERROR_TYPE, MESSAGE_TYPE, STATUS_CODE } = require('../constants/errors');
 
@@ -28,8 +29,22 @@ module.exports.createCard = (req, res) => {
 }
 
 module.exports.deleteCard = (req, res) => {
+  // Сначала проверим переданный cardId на валидность. Если он невалидный, не будем ничего делать
+  const _id = req.params.cardId;
+  const isValidId = mongoose.Types.ObjectId.isValid(_id);
+  if (!isValidId) {
+    return res.status(STATUS_CODE.badRequest).send({ message: MESSAGE_TYPE.cast });
+  }
+
   Card.findByIdAndRemove(req.params.cardId).select('name link owner likes _id')
-    .then(card => res.send({ data: card }))
+    .then(card => {
+      if (!card) {
+        return res.status(STATUS_CODE.notFound).send({ message: MESSAGE_TYPE.noCard });
+      }
+      else {
+        res.send({ data: card });
+      }
+    })
     .catch(err => {
       if (err.name === ERROR_TYPE.cast) {
         return res.status(STATUS_CODE.notFound).send({ message: MESSAGE_TYPE.noCard });
@@ -40,12 +55,26 @@ module.exports.deleteCard = (req, res) => {
 }
 
 module.exports.likeCard = (req, res) => {
+  // Сначала проверим переданный cardId на валидность. Если он невалидный, не будем ничего делать
+  const _id = req.params.cardId;
+  const isValidId = mongoose.Types.ObjectId.isValid(_id);
+  if (!isValidId) {
+    return res.status(STATUS_CODE.badRequest).send({ message: MESSAGE_TYPE.cast });
+  }
+
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } }, // добавить _id в массив, если его там нет
     { new: true },
   ).select('name link owner likes _id')
-    .then(card => res.send({ data: card }))
+    .then(card => {
+      if (!card) {
+        return res.status(STATUS_CODE.notFound).send({ message: MESSAGE_TYPE.noCard });
+      }
+      else {
+        res.send({ data: card });
+      }
+    })
     .catch(err => {
       if (err.name === ERROR_TYPE.validity) {
         return res.status(STATUS_CODE.badRequest).send({ message: MESSAGE_TYPE.validity });
@@ -58,12 +87,26 @@ module.exports.likeCard = (req, res) => {
 }
 
 module.exports.dislikeCard = (req, res) => {
+  // Сначала проверим переданный cardId на валидность. Если он невалидный, не будем ничего делать
+  const _id = req.params.cardId;
+  const isValidId = mongoose.Types.ObjectId.isValid(_id);
+  if (!isValidId) {
+    return res.status(STATUS_CODE.badRequest).send({ message: MESSAGE_TYPE.cast });
+  }
+
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } }, // убрать _id из массива
     { new: true },
   ).select('name link owner likes _id')
-    .then(card => res.send({ data: card }))
+    .then(card => {
+      if (!card) {
+        return res.status(STATUS_CODE.notFound).send({ message: MESSAGE_TYPE.noCard });
+      }
+      else {
+        res.send({ data: card });
+      }
+    })
     .catch(err => {
       if (err.name === ERROR_TYPE.validity) {
         return res.status(STATUS_CODE.badRequest).send({ message: MESSAGE_TYPE.validity });
